Refactor types handling, structure definitions
